generator working
